Initial Commit, basic infrastructure.
The phone basics work without handling of DTMF (SendTones), the theme
is pretty barebones at this point.

Supported API:
        - VoiceCall:
                - Methods: Hangup, Answer
                - Signals: PropertyChanged, DisconnectReason
        - VoiceCallManager:
                - Methods: GetCalls, Dial
                - Signals: CallAdded, CallRemoved
        - Manager:
                - Methods: GetModems
                - Signals: ModemAdded, ModemRemoved, PropertyChanged

It's fully hot-pluggable: the daemon, modems and calls can enter and
exit at anytime.

It's remote controllable using DBus. Right now just Activate and Dial
are implemented so other applications can bring up the dialer window
and request a number to be dialed. This is also used to guarantee a
single instance of the application.