history: try to recover by reusing items stored in log.
if application shutdowns with ongoing calls, they will be stored to
EET, on restart it will try to recover items from log by using line_id
and and start time.

For calls that were not completed, this will not work as Call.StarTime
just exist for active or held calls.