keypad: supplementary services may take long, show loading dialog.
We show a dialog without a message until
SupplementaryServices.Initiate() returns.

Dialogs without message will show loading animation.

The dialog is hold hidden a bit before it's shown. The issue is if
SupplementaryServices.Initiate() fails due NotRecognized or
InvalidFormat. Those are fast and we fallback to Dial()... in this
case we delete the popup and the resulting user interface is a quick
show-hide, which is annoying.