gui: add editable entry support to simple popup.
This refactors the internal popup to keep a context and store the
entry and message, as well as a button. The logic is now changed to
evaluate the context whenever there is at least a message or an entry.

Dismiss button also uses the buttons_set() infra, which will delete
previous buttons if any.