callscreen: remove swap and contacts, make 1x4 grid and different swap action.
Bug #24 complains about the useless behavior of "contacts" button,
since it's duplicating the "add call". To solve it we're removing that
button and also the "swap", reducing from 6 to 4 buttons.

The swap action is now available by clicking the held call. It will
activate that call, which should still be obvious.