ofono: only issue "call added" after properties were populated.
This allow us to decide better where to place the call once we get its
creation signal, not having to rely on following "call changed".

_ofono_dial_reply() was also changed so it does not call _call_add()
with NULL properties anymore, instead it creates the call instance,
adds to the hash and register a pending_dial to be dispatched when the
CallAdded signal is captured. This way we have all the properties as
usual.