rc: add nasty tizen work around.
we've evaluated all the options, but Tizen is completely nuts and will
run the window manager (e17) as root, while the applications are executed
as user "app". Enlightenment will create its session bus for user "root",
but nothing is done regarding "app".

Then for tizen builds we use the system bus, always.