Fixes TIVI-978, where touch calibration is off
The dialer has been configured to run at portrait orientation at 720x1280,
but the Toyota homescreen is resizing to 1920x1080, although the app UI is able
to resize to fit the screen, it seems there's a bug that the keyboard events are
unable to sync with the UI that, which results in user will not be able to touch
the bottom part of the dialer.  A quick patch changing the default layout of the
dialer app to landscape mode seems to eliminate this issue.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>